move AssetType to types.ts
