--- src/providers/assets/constants.ts
@@ -1,6 +1,0 @@
-export enum AssetType {
-  COVER = 'cover',
-  AVATAR = 'avatar',
-  THUMBNAIL = 'thumbnail',
-  MEDIA = 'media',
-}
